chore: Update comments script source URL
The code changes update the source URL for the comments script in the post.hbs file. This change is made to use a different URL (https://sredevopsorg.vercel.app/client.js) for the comments script, which is hosted on a different server. This update ensures that the comments functionality is properly integrated with the SREDevOps.org website.